requirements.txt
bcrypt==4.0.0
click==8.0.3
Flask==1.0.4
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.22.0
flask-sqlalchemy==2.3.0
Werkzeug==2.0.2
WTForms==3.0.0
